+++ backend/session/data/__init__.py
@@ -1,0 +1,6 @@
+"""Provide data classes for session, participant, size and position data.
+
+Provides: `SessionData`, `ParticipantData`, `PositionData` and `SizeData` as well as
+factory functions for `SessionData`: `session_data_factory` and `ParticipantData`:
+`participant_data_factory`, and more
+"""
