+++ backend/session/data/participant/participant_summary.py
