+++ backend/users/participant_factory.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+
 if TYPE_CHECKING:
     from modules.experiment import Experiment
 
@@ -9,7 +10,7 @@ from aiortc import RTCSessionDescription
 
 from modules.connection import connection_factory
 from modules.connection_subprocess import connection_subprocess_factory
-from modules.data import ParticipantData
+from session.data.participant import ParticipantData
 from modules.filter_api import FilterAPI
 
 from server import Config
